rename: card 패키지명 supplies로 수정
