+++ src/main/java/blackjack/domain/game/BlackjackGame.java
@@ -1,7 +1,7 @@
 package blackjack.domain.game;
 
-import blackjack.domain.card.Deck;
-import blackjack.domain.card.Hand;
+import blackjack.domain.supplies.Deck;
+import blackjack.domain.supplies.Hand;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Player;
 import blackjack.domain.gamer.Players;
