+++ src/main/java/blackjack/domain/gamer/Player.java
@@ -1,6 +1,6 @@
 package blackjack.domain.gamer;
 
-import blackjack.domain.card.Card;
+import blackjack.domain.supplies.Card;
 
 import java.util.List;
