+++ src/main/java/blackjack/domain/supplies/Card.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.supplies;
 
 public record Card(Rank rank, Suit suit) {
 }
