+++ src/main/java/blackjack/domain/supplies/Hand.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.supplies;
 
 import java.util.ArrayList;
 import java.util.List;
