+++ src/main/java/blackjack/domain/supplies/Rank.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.supplies;
 
 import java.util.List;
